feature: check return of healthcheck is as expected
the redis-cli will return with an exit code of 0 even if the node is loading and not accepting connections
(responding with `LOADING Redis is loading the dataset in memory`).
`PONG` is returned only when the cluster is ready.
Checking for this return ensures that we only count redis nodes as available if they're actually accepting queries.

Signed-off-by: Daniel Goldman <danielgoldman4@gmail.com>