fix: Check return of probes to ensure Redis is returning responses (#1626)
* exit immediately on command failure

since we'll be using 2 commands in our script,
we want to exit early if the command fails

Signed-off-by: Daniel Goldman <danielgoldman4@gmail.com>

* refactor: extract the redisHealthCheck as a separate variable

we'll be embedding this in a subcommand invocation,
so it's good to have it separated here

Signed-off-by: Daniel Goldman <danielgoldman4@gmail.com>

* feature: check return of healthcheck is as expected

the redis-cli will return with an exit code of 0 even if the node is loading and not accepting connections
(responding with `LOADING Redis is loading the dataset in memory`).
`PONG` is returned only when the cluster is ready.
Checking for this return ensures that we only count redis nodes as available if they're actually accepting queries.

Signed-off-by: Daniel Goldman <danielgoldman4@gmail.com>

* test: update the tests

Signed-off-by: Daniel Goldman <danielgoldman4@gmail.com>

* doc: explain why we're verifying return value of probe

Signed-off-by: Daniel Goldman <danielgoldman4@gmail.com>

---------

Signed-off-by: Daniel Goldman <danielgoldman4@gmail.com>